Avoids exit command terminating cli before engine shutdown
* Exit command calls System.exit which does not shutdown the KEngine before
  terminating the JVM.

* KomodoShell
 * Creates an interface from the implementation and renames impl to
   DefaultKomodoShell
 * Provides methods for accessing engine and streams to related classes

* KomodoShellParent
 * Provides generic interface for two different methods of running cli to
   allow exit to be handled in ways appropriate to parent

* WorkspaceStatusImpl
 * Passes in shell rather than inidividual fields to allow ExitCommand to
   call shell.exit
