+++ plugins/org.komodo.eclipse.shell/META-INF/MANIFEST.MF
@@ -14,4 +14,5 @@ Require-Bundle: org.eclipse.ui;bundle-version="3.106.0",
  org.komodo.spi,
  org.komodo.shell,
  org.komodo.utils,
- org.komodo.core
+ org.komodo.core,
+ org.komodo.shell-api
