+++ plugins/org.komodo.shell-api/src/org/komodo/shell/api/KomodoShell.java
@@ -1,0 +1,55 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ * See the COPYRIGHT.txt file distributed with this work for information
+ * regarding copyright ownership.  Some portions may be licensed
+ * to Red Hat, Inc. under one or more contributor license agreements.
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
+ * 02110-1301 USA.
+ */
+package org.komodo.shell.api;
+
+import java.io.InputStream;
+import java.io.PrintStream;
+import org.komodo.core.KEngine;
+import org.komodo.spi.constants.StringConstants;
+
+/**
+ *
+ */
+public interface KomodoShell extends StringConstants {
+
+    /**
+     * @return the shell's komodo engine
+     */
+    KEngine getEngine();
+
+    /**
+     * @return the shell's input stream
+     */
+    InputStream getInputStream();
+
+    /**
+     * @return the shell's output stream
+     */
+    PrintStream getOutputStream();
+
+    /**
+     * Exit the shell
+     */
+    void exit();
+    
+
+}
