+++ tests/org.komodo.shell.test/META-INF/MANIFEST.MF
@@ -9,4 +9,5 @@ Bundle-Localization: plugin
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Require-Bundle: org.junit;bundle-version="4.11.0",
  org.komodo.spi;bundle-version="[1.0.0,2.0.0)",
- org.komodo.test.utils;bundle-version="[1.0.0,2.0.0)"
+ org.komodo.test.utils;bundle-version="[1.0.0,2.0.0)",
+ org.jboss.tools.locus.mockito
